ai_project_1/ai_project_1/requirements.txt

streamlit
opencv-python
numpy
pillow